Moved ILS data caching support to the base class to make it usable for other drivers too.
