+++ module/VuFind/src/VuFind/ILS/Driver/AbstractBase.php
@@ -26,6 +26,7 @@
  * @link     https://vufind.org/wiki/development:plugins:ils_drivers Wiki
  */
 namespace VuFind\ILS\Driver;
+use Zend\Cache\Storage\StorageInterface;
 
 /**
  * Default ILS driver base class.
@@ -41,11 +42,37 @@ namespace VuFind\ILS\Driver;
 abstract class AbstractBase implements DriverInterface
 {
     /**
+     * Cache for storing ILS data temporarily (e.g. patron blocks)
+     *
+     * @var StorageInterface
+     */
+    protected $cache = null;
+
+    /**
+     * Lifetime of cache (in seconds).
+     *
+     * @var int
+     */
+    protected $cacheLifetime = 30;
+
+    /**
      * Driver configuration
      *
      * @var array
      */
     protected $config = [];
+
+    /**
+     * Set a cache storage object.
+     *
+     * @param StorageInterface $cache Cache storage interface
+     *
+     * @return void
+     */
+    public function setCacheStorage(StorageInterface $cache = null)
+    {
+        $this->cache = $cache;
+    }
 
     /**
      * Set configuration.
@@ -60,5 +87,72 @@ abstract class AbstractBase implements DriverInterface
     public function setConfig($config)
     {
         $this->config = $config;
+    }
+
+    /**
+     * Add instance-specific context to a cache key suffix to ensure that
+     * multiple drivers don't accidentally share values in the cache.
+     * This implementation works anywhere but can be overridden with something more
+     * performant.
+     *
+     * @param string $key Cache key suffix
+     *
+     * @return string
+     */
+    protected function formatCacheKey($key)
+    {
+        return get_class($this) . '-' . md5(json_encode($this->config) . "|$key");
+    }
+
+    /**
+     * Helper function for fetching cached data.
+     * Data is cached for up to $this->cacheLifetime seconds so that it would be
+     * faster to process e.g. requests where multiple calls to the backend are made.
+     *
+     * @param string $key Cache entry key
+     *
+     * @return mixed|null Cached entry or null if not cached or expired
+     */
+    protected function getCachedData($key)
+    {
+        // No cache object, no cached results!
+        if (null === $this->cache) {
+            return null;
+        }
+
+        $fullKey = $this->formatCacheKey($key);
+        $item = $this->cache->getItem($fullKey);
+        if (null !== $item) {
+            // Return value if still valid:
+            if (time() - $item['time'] < $this->cacheLifetime) {
+                return $item['entry'];
+            }
+            // Clear expired item from cache:
+            $this->cache->removeItem($fullKey);
+        }
+        return null;
+    }
+
+    /**
+     * Helper function for storing cached data.
+     * Data is cached for up to $this->cacheLifetime seconds so that it would be
+     * faster to process e.g. requests where multiple calls to the backend are made.
+     *
+     * @param string $key   Cache entry key
+     * @param mixed  $entry Entry to be cached
+     *
+     * @return void
+     */
+    protected function putCachedData($key, $entry)
+    {
+        // Don't write to cache if we don't have a cache!
+        if (null === $this->cache) {
+            return;
+        }
+        $item = [
+            'time' => time(),
+            'entry' => $entry
+        ];
+        $this->cache->addItem($this->formatCacheKey($key), $item);
     }
 }
